fix warning (conversion from 'uint64_t' to 'std::size_t'), Issue #65, and update test program to not require strdup.

+++ sparsepp/spp_utils.h
@@ -195,7 +195,7 @@ inline size_t spp_mix_64(uint64_t a)
     a = (a + (a << 2)) + (a << 4); // a * 21
     a = a ^ (a >> 28);
     a = a + (a << 31);
-    return a;
+    return static_cast<size_t>(a);
 }
 
 template<class ArgumentType, class ResultType>
